Swap order of mmap operations in alloc_chunk for IMA
The Linux Integrity Measurement Architecture (IMA) validates the
integrity of files with hashes and cryptographic signatures. IMA
calculates hash values of executables and other system files at runtime
when they are first accessed.  From an IMA perspective attempting to
execute a file in policy that is currently open for writing makes no
sense and is blocked.

The sljit allocator uses unnamed temp files or memfd to allocate memory
chunks.  The same chunk is mmaped twice once with `PROT_READ | PROT_WRITE`
and once with `PROT_READ | PROT_EXEC`.  IMA will block mmaping a file as
PROT_EXEC if it is already mapped as PROT_WRITE.  This swaps the order
so we first perform the PROT_EXEC mapping.

Since IMA measure and appraises files when they are first accessed,
swapping the order is a legitimate fix.  IMA does not prevent against
memory attacks and is not meant to prevent something like a run-time
alteration of a memory mapped file.

https://ima-doc.readthedocs.io/en/latest/ima-concepts.html